develop

+++ model/basics.go
@@ -113,7 +113,7 @@ type EventItem struct {
 	Sku         string  `json:"sku" gorm:"sku" form:"sku"`
 	Barcode     string  `json:"barcode" gorm:"barcode" form:"barcode"`
 	RetailPrice int     `json:"retail_price" gorm:"retail_price" form:"retail_price"`
-	SalePrice   int     `json:"sale_price" gorm:"sale_price" form:"sale_price"`
+	SalePrice   float32 `json:"sale_price" gorm:"sale_price" form:"sale_price"`
 	Discount    float32 `json:"discount" gorm:"discount" form:"discount"`
 	Season      string  `json:"season" gorm:"season" form:"season"`
 	Category    string  `json:"category" gorm:"category" form:"category"`
@@ -131,10 +131,10 @@ type SaleRecord struct {
 	ItemId     uint64    `json:"item_id" gorm:"item_id" form:"item_id"`
 	OrderId    string    `json:"order_id" gorm:"order_id" form:"order_id"`
 	OrderTime  time.Time `json:"order_time" gorm:"order_time" form:"order_time"`
-	CouponUsed int       `json:"coupon_used" gorm:"coupon_used" form:"coupon_used"`
+	CouponUsed float32   `json:"coupon_used" gorm:"coupon_used" form:"coupon_used"`
 	Source     int       `json:"source" gorm:"source" form:"source"`
 	Quantity   int       `json:"quantity" gorm:"quantity" form:"quantity"`
-	PaidPrice  int       `json:"paid_price" gorm:"paid_price" form:"paid_price"`
+	PaidPrice  float32   `json:"paid_price" gorm:"paid_price" form:"paid_price"`
 	IsReturn   int       `json:"is_return" gorm:"is_return" form:"is_return"`
 	//CreateTime time.Time `json:"create_time" gorm:"create_time" form:"create_time"`
 	//ModifyTime time.Time `json:"modify_time" gorm:"modify_time" form:"modify_time"`
